# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Trade API
  version: 1.0.0
  description: Sandbox crypto trading API for students. Covers market data, wallets, buy, sell, swap, transactions, KYC, and admin operations.
servers:
  - url: /
    description: Current API host
  - url: http://localhost:4200
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, phone, password]
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  example: student@cryptoclass.test
                phone:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User and token
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AuthResponse"
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginType, identifier, password]
              properties:
                loginType:
                  type: string
                  enum: [email, phone]
                  description: Tells the API whether identifier should be matched against email or phone.
                  example: email
                identifier:
                  type: string
                  description: Email address when loginType is email, or phone number when loginType is phone.
                  example: student@cryptoclass.test
                password:
                  type: string
                  example: password123
      responses:
        "200":
          description: User and token, or a 2FA challenge
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AuthResponse"
                twoFactorRequired:
                  $ref: "#/components/examples/TwoFactorChallengeResponse"
...
  /auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Enable authenticator app 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: 2FA enabled
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/TwoFactorEnabledResponse"
  /auth/2fa/verify:
    post:
      tags: [Auth]
      security: []
      summary: Verify a login 2FA challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code]
              properties:
                challengeId:
                  type: string
                  example: 2fa_abc123
                code:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: User and token
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AuthResponse"
  /auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Disable authenticator app 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, code]
              properties:
                password:
                  type: string
                  example: password123
                code:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/TwoFactorDisabledResponse"
  /auth/otp/request:
    post:
      tags: [Auth]
      security: []
      summary: Request demo OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: student@cryptoclass.test
      responses:
        "200":
          description: Demo OTP code
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/OtpRequestResponse"
  /auth/otp/verify:
    post:
      tags: [Auth]
      security: []
      summary: Verify demo OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                  example: student@cryptoclass.test
                code:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/OtpVerifyResponse"
  /auth/kyc:
    post:
      tags: [Auth]
      summary: Submit KYC for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [legalName, country, documentType, documentNumber]
              properties:
                legalName:
                  type: string
                  example: Ada Student
                country:
                  type: string
                  example: Nigeria
                documentType:
                  type: string
                  enum: [national_id, passport, drivers_license]
                  example: national_id
                documentNumber:
                  type: string
                  example: NIN-000-000
                selfieImageUrl:
                  type: string
                  example: https://example.com/uploads/ada-selfie.jpg
                documentImageUrl:
                  type: string
                  example: https://example.com/uploads/ada-national-id.jpg
      responses:
        "201":
          description: KYC submission
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/KycResponse"
  /auth/kyc/uploads:
    post:
      tags: [Auth]
      summary: Create a demo KYC upload URL
      description: Returns a simulated pre-signed upload URL. Mobile apps can upload first, then submit the returned publicUrl in /auth/kyc.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, contentType]
              properties:
                fileName:
                  type: string
                  example: student-national-id.png
                contentType:
                  type: string
                  example: image/png
                documentKind:
                  type: string
                  enum: [selfie, document_front, document_back]
                  example: document_front
      responses:
        "201":
          description: Demo upload instructions
          content:
            application/json:
              examples:
                success:
                  value:
                    data:
                      uploadId: upload_abc123
                      provider: demo_local_storage
                      storageKey: kyc/usr_student/document_front/upload_abc123-student-national-id.png
                      uploadUrl: /storage/uploads/kyc/usr_student/document_front/upload_abc123-student-national-id.png
                      publicUrl: /storage/files/kyc/usr_student/document_front/upload_abc123-student-national-id.png
                      method: PUT
                      headers:
                        content-type: image/png
                      expiresAt: "2026-05-20T13:10:00.000Z"
  /market/assets:
    get:
      tags: [Market]
      security: []
      summary: List supported crypto assets
      parameters:
        - in: query
          name: q
          required: false
          description: Optional search text for asset name or symbol.
          schema:
            type: string
          example: btc
        - in: query
          name: search
          required: false
          description: Alias for q, useful for search inputs.
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
          example: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          example: 10
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [symbol, priceUsd, change24h, minBuyUsd]
          example: priceUsd
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
          example: desc
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AssetListResponse"
  /market/assets/{symbol}:
    get:
      tags: [Market]
      security: []
      summary: Get asset details and demo chart data
      parameters:
        - $ref: "#/components/parameters/Symbol"
      responses:
        "200":
          description: Asset details
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AssetDetailResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Symbol:
      in: path
      name: symbol
      required: true
      schema:
        type: string
      example: BTC
  examples:
...
    AuthResponse:
      summary: Authenticated user
      value:
        data:
          user:
            id: usr_student
            role: customer
            fullName: Ada Student
            email: student@cryptoclass.test
            phone: "+2348010000001"
            twoFactorEnabled: false
            kycStatus: approved
            avatarUrl: null
            watchlist: [BTC, ETH]
            settings:
              language: en
              fiatCurrency: USD
              theme: system
              priceAlerts: true
              pushNotifications: true
              biometricEnabled: false
            createdAt: "2026-05-03T14:00:00.000Z"
          token: demo-user-token
    TwoFactorChallengeResponse:
      summary: 2FA challenge required
      value:
        data:
          requiresTwoFactor: true
          challengeId: 2fa_abc123
          expiresAt: "2026-05-03T14:55:00.000Z"
...
    TwoFactorEnabledResponse:
      summary: 2FA enabled
      value:
        data:
          enabled: true
    TwoFactorDisabledResponse:
      summary: 2FA disabled
      value:
        data:
          enabled: false
    OtpRequestResponse:
      summary: Demo OTP sent
      value:
        data:
          message: Demo OTP sent.
          demoCode: "123456"
          expiresInSeconds: 300
    OtpVerifyResponse:
      summary: OTP verified
      value:
        data:
          verified: true
    KycResponse:
      summary: KYC submission
      value:
        data:
          id: kyc_student
          userId: usr_student
          legalName: Ada Student
          country: Nigeria
          documentType: national_id
          documentNumber: NIN-000-000
          selfieImageUrl: https://images.unsplash.com/photo-1494790108377-be9c29b29330?w=240&h=240&fit=crop&crop=faces
          documentImageUrl: https://images.unsplash.com/photo-1586953208448-b95a79798f07?w=480&h=300&fit=crop
          status: pending
          submittedAt: "2026-05-03T14:05:00.000Z"
          reviewedAt: null
          reviewerNote: null
...
    AssetListResponse:
      summary: Asset list
      value:
        data:
          - id: asset_btc
            symbol: BTC
            name: Bitcoin
            network: Bitcoin Testnet
            priceUsd: 64200.5
            change24h: 2.1
            isActive: true
            minBuyUsd: 10
            minSellUsd: 10
            iconUrl: /assets/btc.svg
          - id: asset_eth
            symbol: ETH
            name: Ethereum
            network: Ethereum Sepolia
            priceUsd: 3100.25
            change24h: -0.8
            isActive: true
            minBuyUsd: 10
            minSellUsd: 10
            iconUrl: /assets/eth.svg
        meta:
          count: 2
    AssetDetailResponse:
      summary: Asset details with chart
      value:
        data:
          id: asset_btc
          symbol: BTC
          name: Bitcoin
          network: Bitcoin Testnet
          priceUsd: 64200.5
          change24h: 2.1
          isActive: true
          minBuyUsd: 10
          minSellUsd: 10
          iconUrl: /assets/btc.svg
          chart:
            - time: "09:00"
              priceUsd: 62916.49
            - time: "10:00"
              priceUsd: 64842.51
            - time: "11:00"
              priceUsd: 63879.5
            - time: "12:00"
              priceUsd: 64200.5
...